remove separate component
