+++ stories/components/SDatePicker.01_Playground.story.vue
@@ -1,14 +1,18 @@
 <script setup lang="ts">
-import SDatePicker from 'sefirot/components/SDatePicker.vue'
+import SDatePickerBase from 'sefirot/components/SDatePicker.vue'
+import { now } from 'sefirot/support/Now'
+import { reactive } from 'vue'
 
 const title = 'Components / SDatePicker / 01. Playground'
+
+const selected = reactive({ ...now })
 </script>
 
 <template>
   <Story :title="title" source="Not available" auto-props-disabled>
     <template #default>
       <Board :title="title">
-        <SDatePicker />
+        <SDatePickerBase v-model="selected" />
       </Board>
     </template>
   </Story>
